fix: display value json input
Computing the display value should retain the original value type of
the value. For example if the value is of type string but the contents
of the string resemble a primitive value the some information may get lost.

Given an value list of strings `["2024.20", "2024.0"]` the json string
passed to `parseDisplayValue` would be the the values `2024.20` and
`2024.0`, which when re-parsed as json values will be interpreted as numbers rather than strings. This has the end effect of the display
values `2024.2` and `2024` while the associated values are `2024.20`
and `2024.0`.

To fix this behavior each value is "json stringified" before sent to
`parseDisplayValue`. For the example above the values would be
`"2024.20"` and `"2024.0"`.